+++ src/main/java/com/atwoz/member/application/event/ValidatedLoginEvent.java
@@ -1,4 +1,4 @@
-package com.atwoz.member.domain.auth;
+package com.atwoz.member.application.event;
 
 import com.atwoz.global.event.Event;
 import lombok.Getter;
