+++ src/main/java/com/atwoz/member/domain/auth/JsonMapper.java
@@ -1,6 +1,6 @@
-package com.atwoz.member.domain.oauth;
+package com.atwoz.member.domain.auth;
 
-import com.atwoz.member.infrastructure.oauth.dto.MemberInfo;
+import com.atwoz.member.application.auth.dto.MemberInfo;
 
 public interface JsonMapper {
