+++ src/main/java/com/atwoz/member/domain/auth/OAuthConnectionManager.java
@@ -1,6 +1,6 @@
-package com.atwoz.member.domain.oauth;
+package com.atwoz.member.domain.auth;
 
-import com.atwoz.member.ui.auth.support.oauth.OAuthProvider;
+import com.atwoz.member.infrastructure.auth.dto.OAuthProvider;
 
 public interface OAuthConnectionManager {
